analysis/pyproject.toml
[tool.poetry]
name = "acme-cultural-analysis"
version = "1.0.0"
description = "Enterprise-grade cultural funding analysis pipeline"
authors = ["ACME Data Science Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.0"
numpy = "^1.26.0"
openpyxl = "^3.1.2"
scikit-learn = "^1.4.0"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
plotly = "^5.19.0"
wordcloud = "^1.9.0"
python-docx = "^1.1.0"
pillow = "^10.2.0"
rich = "^13.7.0"
typer = "^0.9.0"
httpx = "^0.26.0"
tenacity = "^8.2.0"
scipy = "^1.12.0"
statsmodels = "^0.14.0"
pydantic-settings = "^2.9.1"
kaleido = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
ipykernel = "^6.29.0"
jupyter = "^1.0.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"